requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.11.16
aiosignal==1.3.2
aiosqlite==0.21.0
annotated-types==0.7.0
anyio==4.9.0
appdirs==1.4.4
APScheduler==3.11.0
asttokens==3.0.0
attrs==25.3.0
Authlib==1.6.4
backoff==2.2.1
banks==2.1.2
bcrypt==4.3.0
beautifulsoup4==4.13.3
blinker==1.9.0
blis==1.3.0
build==1.3.0
cachelib==0.13.0
cachetools==5.5.1
catalogue==2.0.10
certifi==2025.8.3
cffi==1.17.1
cfgv==3.4.0
charset-normalizer==2.0.12
chromadb==1.1.1
click==8.1.8
cloudpathlib==0.21.1
colorama==0.4.6
coloredlogs==15.0.1
confection==0.1.5
contourpy==1.3.2
crewai==1.3.0
cryptography==44.0.2
cycler==0.12.1
cymem==2.0.11
dataclasses-json==0.6.7
decorator==5.2.1
Deprecated==1.2.18
dirtyjson==1.0.8
diskcache==5.6.3
distlib==0.4.0
distro==1.9.0
docstring_parser==0.17.0
durationpy==0.10
easyocr==1.7.2
en_core_web_sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.8.0/en_core_web_sm-3.8.0-py3-none-any.whl#sha256=1932429db727d4bff3deed6b34cfc05df17794f4a52eeb26cf8928f7c1a0fb85
et_xmlfile==2.0.0
executing==2.2.0
faiss-cpu==1.10.0
fake-useragent==2.2.0
filelock==3.17.0
filetype==1.2.0
Flask==3.0.0
Flask-Bcrypt==1.0.1
Flask-Cors==4.0.0
Flask-Mail==0.10.0
Flask-Session==0.8.0
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.2
flatbuffers==25.9.23
fonttools==4.57.0
frozenlist==1.5.0
fsspec==2025.2.0
genai==0.1.0
git-filter-repo==2.47.0
google-ai-generativelanguage==0.6.15
google-api-core==2.24.1
google-api-python-client==2.161.0
google-auth==2.38.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.3
google-generativeai==0.8.5
google_search_results==2.4.2
googleapis-common-protos==1.67.0
greenlet==3.2.3
griffe==1.7.3
grpcio==1.70.0
grpcio-status==1.70.0
gunicorn==23.0.0
h11==0.14.0
httpcore==1.0.7
httplib2==0.22.0
httptools==0.7.1
httpx==0.28.1
httpx-sse==0.4.3
huggingface-hub==0.29.1
humanfriendly==10.0
identify==2.6.15
idna==3.10
imageio==2.37.0
importlib_metadata==8.7.0
importlib_resources==6.5.2
instructor==1.12.0
ipython==9.1.0
ipython_pygments_lexers==1.1.1
itsdangerous==2.2.0
jedi==0.19.2
jellyfish==1.2.0
Jinja2==3.1.5
jiter==0.9.0
joblib==1.4.2
json5==0.12.1
json_repair==0.25.2
jsonpatch==1.33
jsonpointer==3.0.0
jsonref==1.1.0
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
kiwisolver==1.4.8
kubernetes==34.1.0
langchain==1.0.3
langchain-core==1.0.3
langchain-openai==1.0.2
langcodes==3.5.0
langgraph==1.0.2
langgraph-checkpoint==3.0.1
langgraph-prebuilt==1.0.2
langgraph-sdk==0.2.9
langsmith==0.4.41
language_data==1.3.0
lazy_loader==0.4
llama-cloud==0.1.26
llama-cloud-services==0.6.31
llama-index==0.12.41
llama-index-agent-openai==0.4.10
llama-index-cli==0.4.3
llama-index-core==0.12.41
llama-index-embeddings-openai==0.3.1
llama-index-indices-managed-llama-cloud==0.7.6
llama-index-llms-gemini==0.5.0
llama-index-llms-openai==0.4.5
llama-index-multi-modal-llms-openai==0.5.1
llama-index-program-openai==0.3.2
llama-index-question-gen-openai==0.3.1
llama-index-readers-file==0.4.9
llama-index-readers-llama-parse==0.4.0
llama-parse==0.6.31
lxml==5.3.1
marisa-trie==1.3.0
Markdown==3.7
markdown-it-py==4.0.0
MarkupSafe==3.0.2
marshmallow==3.26.1
matplotlib==3.10.1
matplotlib-inline==0.1.7
mcp==1.20.0
mdurl==0.1.2
memory-profiler==0.61.0
mmh3==5.2.0
mpmath==1.3.0
msgspec==0.19.0
multidict==6.3.2
murmurhash==1.0.13
mypy_extensions==1.1.0
nest-asyncio==1.6.0
networkx==3.4.2
ninja==1.11.1.4
nltk==3.9.1
nodeenv==1.9.1
numpy==2.2.3
oauthlib==3.3.1
onnxruntime==1.23.2
openai==1.109.1
opencv-python-headless==4.11.0.86
openpyxl==3.1.5
opentelemetry-api==1.38.0
opentelemetry-exporter-otlp-proto-common==1.38.0
opentelemetry-exporter-otlp-proto-grpc==1.38.0
opentelemetry-exporter-otlp-proto-http==1.38.0
opentelemetry-proto==1.38.0
opentelemetry-sdk==1.38.0
opentelemetry-semantic-conventions==0.59b0
orjson==3.11.4
ormsgpack==1.12.0
outcome==1.3.0.post0
overrides==7.7.0
packaging==24.2
pandas==2.2.3
pandas-stubs==2.2.3.250308
parso==0.8.4
pdfminer.six==20250327
pdfplumber==0.11.6
pillow==10.4.0
platformdirs==4.3.8
portalocker==2.7.0
posthog==5.4.0
praw==7.8.1
prawcore==2.4.0
pre_commit==4.3.0
preshed==3.0.10
prompt_toolkit==3.0.50
propcache==0.3.1
proto-plus==1.26.0
protobuf==5.29.3
psutil==7.0.0
psycopg2==2.9.10
psycopg2-binary==2.9.10
pure_eval==0.2.3
pyasn1==0.6.1
pyasn1_modules==0.4.1
pybase64==1.4.2
pyclipper==1.3.0.post6
pycparser==2.22
pydantic==2.12.4
pydantic-settings==2.11.0
pydantic_core==2.41.5
Pygments==2.19.1
PyJWT==2.10.1
PyMuPDF==1.26.4
pyparsing==3.2.1
pypdf==5.6.0
PyPDF2==3.0.1
pypdfium2==4.30.1
PyPika==0.48.9
pyproject_hooks==1.2.0
pyreadline3==3.5.4
PySocks==1.7.1
pytesseract==0.3.13
python-bidi==0.6.6
python-dateutil==2.9.0.post0
python-docx==1.1.2
python-dotenv==1.1.1
python-multipart==0.0.20
pytz==2025.1
pywin32==311
PyYAML==6.0.2
RapidFuzz==3.14.0
referencing==0.36.2
regex==2024.11.6
requests==2.32.4
requests-oauthlib==2.0.0
requests-toolbelt==1.0.0
rich==14.1.0
rpds-py==0.24.0
rsa==4.9
safetensors==0.5.3
schedule==1.2.2
scikit-image==0.25.2
scikit-learn==1.6.1
scipy==1.15.2
selenium==4.35.0
sentence-transformers==3.4.1
setuptools==75.8.2
shapely==2.1.0
shellingham==1.5.4
six==1.17.0
skillNer==1.0.3
smart_open==7.3.0.post1
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.6
spacy==3.8.7
spacy-legacy==3.0.12
spacy-loggers==1.0.5
SQLAlchemy==2.0.41
srsly==2.5.1
sse-starlette==3.0.3
stack-data==0.6.3
starlette==0.50.0
striprtf==0.0.26
sympy==1.14.0
tenacity==9.1.2
thinc==8.3.6
threadpoolctl==3.5.0
tifffile==2025.3.30
tiktoken==0.9.0
tokenizers==0.21.0
tomli==2.3.0
tomli_w==1.2.0
torch==2.7.0
torchvision==0.22.0
tqdm==4.67.1
traitlets==5.14.3
transformers==4.49.0
trio==0.30.0
trio-websocket==0.12.2
typer==0.16.1
types-pytz==2025.2.0.20250326
typing-inspect==0.9.0
typing-inspection==0.4.2
typing_extensions==4.14.1
tzdata==2025.1
tzlocal==5.3.1
Unidecode==1.4.0
update-checker==0.18.0
uritemplate==4.1.1
urllib3==2.3.0
uv==0.9.7
uvicorn==0.38.0
vdom==0.6
virtualenv==20.34.0
wasabi==1.1.3
watchfiles==1.1.1
wcwidth==0.2.13
weasel==0.4.1
websocket-client==1.8.0
websockets==15.0.1
Werkzeug==3.1.3
Wikipedia-API==0.8.1
wrapt==1.17.2
wsproto==1.2.0
WTForms==3.2.1
xxhash==3.6.0
yarl==1.19.0
zipp==3.23.0
zstandard==0.25.0
